renamed TripInfo to TripOpen and it required some name changes throughout the application
